core/quorum: check if transaction was already applied
Due to the asynchronous eventing system its possible that transactions
are applied on the pending state multiple times. Once with the TxPreEvent
event and once as a result of the ChainHeadEvent that will fetch all
processable transaction from the transaction pool. This leads to an
invalid nonce message.

This PR will keep track of transactions applied to the pending state
and will verify that an transaction was not applied before.