Poll chain state for contract callback invocation
Previously we relied on filters to watch for block events for firing contract-
creation callback functions. With this change, we now poll the chain's state on
the javascript side using `setInterval`.

This is necessary for low-latency consensus mechanisms like Raft, where a block
event reguarly fires before the web3 layer is able to set the filter to watch
for such events.

This fixes #86.