Connect to initial peers up-front
We now use ConfState to tell us which peers to save in the snapshot,
which frees us from having extra logic in the form of `withoutSelf`.

This also cleans up initial snapshotting that used to occur when each of
the initial conf entries was applied -- we would include the current
node in all snapshots, even though its entry might not have occurred
yet, and there would be only a subset of addresses in the snapshots
until the last node joined, which was inconsistent with the ConfState's
view that all nodes already joined the cluster. Now we connect to all
initial peers up-front, and ignore their "expected" conf change entries
when they occur.