Adding signer tests for EIP155, Homestead and private tx.
In Quorum, private transactions are marked by v=37 and v=38,
this means the chainId is not set in the signed transcation param
`v` as it is in EIP155Singer. This also mean chainId 1 cannot be
used, as it will conflict with a public transaction signed by
EIP155Signer, `v=chainId * 2 + 35` and `v=chainId * 2 + 36`
which for `chainId = 1` would be 37 and 38.
Test uses 2 fixed keys which will resolve to the recovery bit being
set to 0 and 1.
1. signing and recovering public txs with EIP155Singer and Homestead.
2. demonstrating the conflict when chainId is set to 1.
3. signing a private transaction with Homestead and recoverying with
   EIP155Signer.
4. signing and recovering private txs with Homestead.
5. signing and recovery failure when signed with EIP155.