Updates quickstart to highlight quorum-examples
The link to quorum-examples and the explanation that it is the recommended way to get started was kind of lost in the middle of a long intro paragraph. This made my eyes jump to Building Quorum Node from Source, thinking it was the first step in getting started. Adding an explicit header and section for Quorum Examples makes it impossible to miss.