VM read-only mode: check `operation.writes` not `op.isMutating`.
`op.isMutating()` was added for checking mutating VM operation, but `operation.writes`
should be used now as it registers all mutating ops and is consistent with upstream.
Additionally, there is a new  mutating opcode `CREATE2`, but `CREATE2` has not been added to the
`op.isMutating()` set. The VM is in read-only mode when a private contract calls a public contract.