parent c215989c10f191924e6b5b668ba4ed8ed425ded1 author vsmk98 <vsaimk@gmail.com> 1585565614 +0800 committer Trung Nguyen <24930+trung@users.noreply.github.com> 1588174146 -0400 gpgsig -----BEGIN PGP SIGNATURE-----
 iQEzBAABCAAdFiEEzAFy365WU5uP2ESgRjZDTtlQXrcFAl6pnUIACgkQRjZDTtlQ
 XrcL2Af+IKHLFkPhSoEHvhN+KDEKqNt1GFZQuLLUhgP9yFsbnbgPlUaYPSkWTAOk
 v0EMUnYUjNsKSj6PsWj4CaW2DBovInIGE/R1oWjmpjPY+IiLJZHg9S+3o9D/pRd6
 a/jXcTdsL9t/u9RqZICzJwVEWcSFNHWH8/EhWpf/NExrxgiue9Ib146hzJn7Rzss
 vgts6vco87r+ck6SZlqNrldwz/V1kKQXDRsTrPC4h7sJ6DTG+NgGDjXu9A5PuVB2
 9byGG1EDNTuqhaHJoUu32m3qs61WKXprTAuo3Bo4i38yemiotPKotBCNKo1odbrV
 h9mseIdlSfWEOx0SdfbVXe3qMMYe7w==
 =7LJs
 -----END PGP SIGNATURE-----

Enable networks to change maxCodeSize multiple times #975

Current maxCodeSize implementation allows change of maxCodeSize once in network life.
If the maxCodeSize is changed multiple times, its possible that any new node joining
the network can get Bad block error while syncing the blocks as the historical value
of maxCodeSize at that particular block height may not match with the current maxCodeSize.
It should be possible to track the historical changes of maxCodeSize in genesis and
thus allow it to be changed multiple times in network life.