Make raftId deterministic, uint64 based on enodeId.
Change raftId from uint16 -> uint16. uint64 is the underlying representation
of the raftId in etcd. The raft id can then be changed from a monotonically
increasing uint16 to an id that is derived from node's enode id. This allows
for a determininistic raftId, allowing nodes to be added and removed with the
same raftId. This simplifies the add node requirement of having to obtain the
raftId from an operator and changing the startup param to reflect this
each time it joinsthe network, e.g. a node is added/removed.

* Truncate enodeID so that it fits inside a uint64, this fixes the issue betweent
raft protocol id uint64 and uint16 diff, and by making the raftId on quorum
side uint64 to match raft protocol side we have a larger raftId space to avoid
collisions.
TODO: hash the shortened enodeId with a cryptographic hash to help ensure uniquenesss.
* Remove initializing the Raft ProtocolManager with the raftId, and
calculating the peer raftIds based on their order inthe static-nodes.json file,
these raftId are now determined when the raft protocol manager is started up
and are not dependent on order, but only on the enodeId of each nodes.
* Change snapshoting removedpeer map from uint16->uint64, the removedpeer is
no longer permenent as there is no risk/extremely low risk of two raft nodes
obtaining the same raftId, as they are generated from the unique enode ids.
* The removedPeers map is kept, but nodes can now be re-added with the same
Id after they were deleted, to enable this remove the re-added peer from the
removedPeer map when proposing the node and when receiving and AddNode config
change.
* Tested, 3 node cluster, add node4, delete node4 readd node4.

Update raft api to accomidate raftId changes.

* raftId is now a uint64, in order to display this properly in Javascript, the
Javascript representation must be a string.
* Update the raft api functions to reflect the raftId changes.
* Make the raftId 64 bits.
* Tested with, add / remove / readd node, and add learner / promote.
* Add support for raft removePeer via nodeId.