Remove --raftjoinexisting flag.
This is no longer needed as nodes know their own raftId based on their
enodeid, so do not have to be passed in their raft id obtained from a
network operator or cluster member. The static-nodes.json file is used
to get the initial set of nodes that raft knows about. If the
node is joining a cluster, it will use its static-nodes.json include any
nodes that it knows about there, and will learn about the other nodes from the
cluster. The new node no longer has to specify its raftid at startup as it generates
the id from its enode id, in the same way the node that added it does.