[raft] fix double addPeer, should also check self.
When checking if an enode is already in the raft cluster, the node should
also check if it is adding its own enode.

This fixes the case where a node can be added twice by running `raft.addPeer(enodeURL)`
on itself, e.g. running on a node with the same enodeURL as the node being added.
The node will allow the enode to be added again, after which the cluster will be in a
weird state containing two peers with the same enodeURL, the duplicate peer will
be in a bad state and will not be conneted.

Fix by adding a check in `isNodeAlreadyInClusteri(enode)` to check if the enode is
the enode of running node, not just one of the peers.

In `minter_test.go` the p2p.Server is now mocked out when creating the
RaftService, this is necessary as `isNodeAlreadyInCluster` uses the
p2p.Server to check if this is the calling node (itself). Mocking out
the p2p.Server avoids a 'invalid memory address or
nil pointer dereference' exception thrown in the `minter_test.go`.