introduce AccountExtraData
PrivacyMetadata was added in Privacy Enhancements. It is linked to each
private contract account and stored in a trie to allow rollback,
especially for contract extension.

This change is to generalize the extra data that supplements the account
data by introducing AccountExtraData. The first extra data is
PrivacyMetadata.

An additional improvement is to make sure AccountExtraData is also
linked to public accounts for future needs.