add eth service support for multiple subprotocols.
Enable the eth service to support multiple subprotocols, e.g.
"eth" AND "istanbul/100", instead of a single protocol, e.g.
"eth" OR "istanbul/99".

When istanbul consensus was originally added, the eth service was
modified to override the eth subprotocol maintained by the eth service
to a quorum specific consensus protocol, e.g. "istanbul/99" meaning all
eth p2p messages would be sent over a consensus specific subprotocol in
addition to any consensus protocol message. In the case of istanbul,
the "eth" subprotocol would no long be included as a devp2p capability/
subprotocol for the node.

This commit changes that, and enables the eth service to return multiple
subprotocols, e.g. "eth" and "istanbul/100".

"istanbul/100" has been added as a new istanbul subprotocol version,
which supports the multple subprotocol design. peers running older
versions of istanbul will continue to override the "eth" service
and will communicate over a single subprotocol:

old node: istanbul/99 <-> istanbul/99
new node: istanbul/100, eth <-> istanbul/100, eth

To maintain backwards compatibility, when establishing a connection with a
new peer that does not support mutliple protocols, but rather a "legacy"
protocol, e.g. "istanbul/64" "istanbul/99" "clique" (v2.6), the nodes
will communicate over the single "legacy" protcol and not "eth" + consensus
subprotocol.

When running two subprotcols, "eth" + "istanbul" (consensus), the "eth"
subprotocol manages the lifecycle of the peer for the eth service, and
the consensus subprotocol uses this peer for its communication, registering
a separate protocol read writer on it.

To support this, a channel is added to a p2p peer (EthPeerRegistered), which
is used to communicate to the "consensus" / quorum subprotocols that the
eth peer is ready and can be retrieved.

When running a quorum specific subprotocol on the eth service, the eth
peer is started and managed by the "eth" subprotocol. The quorum specific
subprotocol uses this peer to send messages.

This channel is necessary, as every protocol registerd in a p2p peer
running map, the Run method is called inside a goroutine. When iterating
through the running map,  the order is not guaranteed and inside the
goroutines we cannot count on execution time of the methods. Thus, we
use the EthPeerRegistered channel to ensure that the eth protocol has
register the peer before attempting to look it up.

maintains backwards compatiblity with: istanbul, raft, clique.