eth: fix issue peer <=eth64 unsupported  message.
eth64, eth63 do not support NewPooledTransactionHashesMsg, as it
was added with eth65. If a peer running eth version <=eth65 receives
NewPooledTransactionHashesMsg the peer will disconnect as it is
an unsupported message. NewPooledTransactionHashesMsg is used
in p.announceTransactions(). Only run announceTransaction() if
the peer we are registering is running version  >=eth65.