Update web3.js used by geth console to check PMT's internal private tx for contract addresses
This resolves an issue when providing a callback to the contract constructor when using web3js abi (e.g. simpleContract.new(..., [callback]).  The callback is called multiple times: once when the transaction is sent and at least one more time when the transaction is minted and the contract is created.

To check the contract is created, web3.js checks if there is a contractAddress in the transactionReceipt.  In the case of PMT, the transactionReceipt (for the PMT) will always have null contractAddress.  As a result, the user cannot retrieve the created contractAddress using the constructor callback.  Additionally, the user will also receive an error in the js console after 50 blocks ("err creating contract Error: Contract transaction couldn't be found after 50 blocks").

This could be misleading and inconvenient for users so this commit also checks the privateTransactionReceipt for a contract address to cover scenarios where PMTs are being used.  This allows the user to access the created contract address using the constructor callback, and prevents the error message after 50 blocks.

go generate has been run in the internal/jsre/deps package to update bindata.go which contains a static store of the zipped binary of web3.js.